warning fix

+++ linux/drm_lists.h
@@ -214,7 +214,8 @@ int DRM(freelist_put)(drm_device_t *dev, drm_freelist_t *bl, drm_buf_t *buf)
 	atomic_inc(&bl->count);
 	if (atomic_read(&bl->count) > dma->buf_count) {
 		DRM_ERROR("%ld of %d buffers free after addition of %d\n",
-			  atomic_read(&bl->count), dma->buf_count, buf->idx);
+			  (unsigned long)atomic_read(&bl->count), 
+				dma->buf_count, buf->idx);
 		return 1;
 	}
 				/* Check for high water mark */
